REFACTOR: Introduce RaiseError
I as trying to read throught the code to grok things and felt
quite distracted by the error code all over the place so I
decided to refactor it into a common component. Another possible
refactoring is to remove the magic strings and move to properties
or text enums to reduce possible errors and future mistakes.